chore(fileimport): remove unused parser.js file
