+++ packages/fileimport-service/ifc/index.js
@@ -1,6 +1,6 @@
 const { performance } = require('perf_hooks')
 const { fetch } = require('undici')
-const Parser = require('./parser_v2')
+const Parser = require('./parser')
 const ServerAPI = require('./api.js')
 const Observability = require('@speckle/shared/dist/commonjs/observability/index.js')
 const { logger: parentLogger } = require('../observability/logging')
